Refactor initial tutorial
