+++ docker-compose.yml
@@ -44,4 +44,4 @@ services:
       MYSQL_PASSWORD: "spring123"
       MYSQL_ROOT_PASSWORD: "root666"
     ports:
-    - "3306"
+    - "3306:3306"
