+++ pan-pan-backend/src/main/java/panpan/PanPanApplication.java
@@ -1,7 +1,10 @@
-package com.desenho.panpan;
+package panpan;
+
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+
 
 @SpringBootApplication
 public class PanPanApplication {
