+++ pan-pan-backend/src/main/java/panpan/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.desenho.panpan.controller;
+package panpan.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
