+++ pan-pan-backend/src/main/java/panpan/model/User.java
@@ -1,16 +1,22 @@
-package com.desenho.panpan.model;
+package panpan.model;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-
-import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.Date;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.Data;
 
 @Data
 @Entity
 public class User {
 
+	@Id
+	@GeneratedValue
     private Long id;
     private String firstName;
     private String lastName;
@@ -21,17 +27,15 @@ public class User {
     private String state;
     private String city;
     private String role;
-
-    @ElementCollection
     private ArrayList<String> instruments;
-
+    
     @JsonProperty
     private Boolean active;
 
     public User() {
         super();
-        setActive(Boolean.FALSE);
-        setRole("User");
+        active = (Boolean.FALSE);
+        role = ("User");
     }
 
     public User(String email, String password){
