+++ pan-pan-backend/src/test/java/panpan/PanPanApplicationTests.java
@@ -1,4 +1,4 @@
-package com.desenho.panpan;
+package panpan;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
